CoachNutrition v0.1.1

+++ CoachNutritionApp/app/src/main/java/groupk/coachnutrition/DayShowActivity.java
@@ -3,29 +3,37 @@ package groupk.coachnutrition;
 import android.app.LoaderManager;
 import android.content.ContentUris;
 import android.content.CursorLoader;
+import android.content.DialogInterface;
 import android.content.Loader;
 import android.database.Cursor;
 import android.net.Uri;
 import android.os.Bundle;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
+import android.widget.Button;
 import android.widget.EditText;
 import android.widget.SimpleCursorAdapter;
 import android.widget.Spinner;
 import android.widget.TextView;
 
+import intermediatecontentprovider.IntermediateCoachNutrition;
 import modules.History;
 import modules.HistoryMeal;
 import modules.Meal;
-import intermediatecontentprovider.IntermediateCoachNutrition;
 
 /**
+ * Group K
+ * 
+ * @author Slimane SIROUKANE
+ * @author Fatima CHIKH
+ * 
  * Day Show Activity to display different information about selected History
  */
-public class DayShowActivity extends AppCompatActivity implements LoaderManager.LoaderCallbacks<Cursor>{
+public class DayShowActivity extends AppCompatActivity implements View.OnClickListener, LoaderManager.LoaderCallbacks<Cursor>{
     public String day_date;
     public IntermediateCoachNutrition inter;
     public float day_min;
@@ -33,7 +41,7 @@ public class DayShowActivity extends AppCompatActivity implements LoaderManager.
     public History h;
     public HistoryMeal hm;
     public int   idh;
-    public int idm;
+    public boolean submit_delete;
     public Spinner spinner_day_show;
     public SimpleCursorAdapter adapter_day_show;
     public LoaderManager manager_day_show;
@@ -100,6 +108,14 @@ public class DayShowActivity extends AppCompatActivity implements LoaderManager.
             }
         });
 
+
+        Button b_update_meal = (Button) findViewById(R.id.b_update_day_meal);
+        b_update_meal.setOnClickListener(this);
+
+
+        Button b_delete_meal = (Button) findViewById(R.id.b_delete_day_meal);
+        b_delete_meal.setOnClickListener(this);
+
         /****manager_day_show****/
         manager_day_show = getLoaderManager();
         manager_day_show.initLoader(0, null, this);
@@ -111,7 +127,8 @@ public class DayShowActivity extends AppCompatActivity implements LoaderManager.
      * 
      * @param view
      */
-    public void dayMealOnClick(View view){
+    @Override
+    public void onClick(View view){
         //id of selection in adapter_day_meal
         if(spinner_day_show != null){
             int position = spinner_day_show.getSelectedItemPosition();
@@ -126,31 +143,60 @@ public class DayShowActivity extends AppCompatActivity implements LoaderManager.
             case R.id.b_update_day_meal:
                 Log.i("b_update_day_meal", "b_update_day_meal button get clicked");
                 // set new quantity on history meal
-                hm.setQuantity(Float.parseFloat(((EditText) findViewById(R.id.et_meal_quantity)).getText().toString()));
-                //Get old Quantity in HistoryMeal and TotalCalories
-                float old_TotalCaloriesHM = hm.getTotalCaloriesHM();
-                //Get Calories
-                int idm = hm.getIdMeal();
-                Meal m = inter.getMealById(idm);
-                Log.d("dayShow","calories " + m.getCalorie());
-                //Update new TotalCalories
-                hm.setTotalCaloriesHM((m.getCalorie()*hm.getQuantity()));
-                inter.updateHistoryMeal(hm);
-                //Update TotalCaloriesH in History
-                h.setTotalCal(h.getTotalCal() + hm.getTotalCaloriesHM()-old_TotalCaloriesHM);
-                inter.updateHitory(h);
-                ((TextView) findViewById(R.id.tv_day_total_cal)).setText(h.getTotalCal()+"");
+
+                String quantity_s = ((EditText) findViewById(R.id.et_meal_quantity)).getText().toString();
+                if(!quantity_s.isEmpty()) {
+                    hm.setQuantity(Float.parseFloat(quantity_s));
+                    //Get old Quantity in HistoryMeal and TotalCalories
+                    float old_TotalCaloriesHM = hm.getTotalCaloriesHM();
+                    //Get Calories
+                    int idm = hm.getIdMeal();
+                    Meal m = inter.getMealById(idm);
+                    Log.d("dayShow", "calories " + m.getCalorie());
+                    //Update new TotalCalories
+                    hm.setTotalCaloriesHM((m.getCalorie() * hm.getQuantity()));
+                    inter.updateHistoryMeal(hm);
+                    //Update TotalCaloriesH in History
+                    h.setTotalCal(h.getTotalCal() + hm.getTotalCaloriesHM() - old_TotalCaloriesHM);
+                    inter.updateHitory(h);
+                    ((TextView) findViewById(R.id.tv_day_total_cal)).setText(h.getTotalCal() + "");
+                }else{
+                    IntermediateCoachNutrition.notification(this, "Please, insert a quantity", 1);
+                }
                 break;
 
             case R.id.b_delete_day_meal:
                 Log.i("b_delete_day_meal", "b_delete_day_meal button get clicked");
-                //update TotalCalories History
-                h.setTotalCal(h.getTotalCal()-hm.getTotalCaloriesHM());
-                //delete MealHistoryMeal
-                inter.deleteMealHistoryMeal(hm.getIdMeal(),hm.getIdHistory());
-                //Update TotalCalorie History
-                inter.updateHitory(h);
-                ((TextView) findViewById(R.id.tv_day_total_cal)).setText(h.getTotalCal()+"");
+                DialogInterface.OnClickListener dialogClickListener = new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        switch (which){
+                            case DialogInterface.BUTTON_POSITIVE:
+                                submit_delete = true;
+                                break;
+
+                            case DialogInterface.BUTTON_NEGATIVE:
+                                submit_delete = false;
+                                break;
+                        }
+                    }
+                };
+
+                AlertDialog.Builder builder2 = new AlertDialog.Builder(this);
+                builder2.setMessage("Are you sure you want to delete this?")
+                        .setPositiveButton("Yes", dialogClickListener)
+                        .setNegativeButton("No", dialogClickListener)
+                        .show();
+
+                if(submit_delete) {
+                    //update TotalCalories History
+                    h.setTotalCal(h.getTotalCal() - hm.getTotalCaloriesHM());
+                    //delete MealHistoryMeal
+                    inter.deleteMealHistoryMeal(hm.getIdMeal(), hm.getIdHistory());
+                    //Update TotalCalorie History
+                    inter.updateHitory(h);
+                    ((TextView) findViewById(R.id.tv_day_total_cal)).setText(h.getTotalCal() + "");
+                }
                 break;
 
             default:
